server/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
aiosqlite==0.19.0
alembic==1.13.0
annotated-types==0.6.0
anyio==3.7.1
apprise==1.6.0
asgi-lifespan==2.1.0
asgiref==3.7.2
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
coolname==2.2.0
croniter==2.0.1
cryptography==41.0.7
dateparser==1.2.0
Deprecated==1.2.14
dnspython==2.4.2
docker==6.1.3
elastic-transport==8.11.0
elasticsearch==8.11.1
email-validator==2.1.0.post1
fastapi==0.105.0
fastapi-utils==0.2.1
frozenlist==1.4.1
fsspec==2023.12.2
google-auth==2.25.2
googleapis-common-protos==1.59.1
graphviz==0.20.1
greenlet==3.0.2
griffe==0.38.1
grpcio==1.60.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.25.2
hyperframe==6.0.1
idna==3.6
importlib-metadata==6.11.0
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kubernetes==28.1.0
loguru==0.7.2
Mako==1.3.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
oauthlib==3.2.2
opentelemetry-api==1.21.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-asgi==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
opentelemetry-util-http==0.42b0
orjson==3.9.10
packaging==23.2
pathspec==0.12.1
pendulum==2.1.2
prefect==2.14.11
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.13
pydantic_core==2.14.5
Pygments==2.17.2
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
readchar==4.0.5
redis==5.0.1
referencing==0.32.0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.15.2
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
schedule==1.2.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.50
starlette==0.27.0
temporalio==1.4.0
tenacity==8.2.3
text-unidecode==1.3
thrift==0.16.0
toml==0.10.2
typer==0.9.0
types-protobuf==4.24.0.4
typing_extensions==4.9.0
tzlocal==5.2
ujson==5.9.0
urllib3==1.26.18
uvicorn==0.24.0.post1
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
